move run to lib
